Make Grid3D a template
