+++ src/containers/grid_3d.h
@@ -7,17 +7,18 @@
 
 namespace containers {
 
+    template<class T, T default_value>
     class Grid3D {
     private:
         int width = 1;
         int depth = 1;
         int height = 1;
 
-        Vector<Object*> internal_array;
+        Vector<T> internal_array;
 
         bool is_in_range(int x, int y, int z) const;
 
-        int get_id(const Vector3 &position) const;
+        int get_index_from_position(const Vector3 &position) const;
 
         Vector3 get_position_3d(int id) const;
 
@@ -38,17 +39,17 @@ namespace containers {
 
         void reset();
 
-        Object* get_data(const Vector3 &position);
+        T get_data(const Vector3 &position);
 
-        void set_data(const Vector3 &position, Object* data);
+        void set_data(const Vector3 &position, T data);
 
-        bool insert_box(const AABB &aabb, Object* data, bool remove = false);
+        bool insert_box(const AABB &aabb, T data, bool remove = false);
 
         bool is_overlapping(const AABB &aabb) const;
 
-        bool has(Object* object) const;
+        bool has(T object) const;
 
-        void for_each(void (* fptr)(Object*));
+        void for_each(void (* fptr)(T));
 
         int get_width() const;
 
@@ -62,10 +63,217 @@ namespace containers {
 
         void set_height(int h);
 
-        const Vector<Object*> &get_internal_array() const;
+        const Vector<T> &get_internal_array() const;
 
-        void set_internal_array(const Vector<Object*> &array);
+        void set_internal_array(const Vector<T> &array);
     };
+
+    template<class T, T default_value>
+    void Grid3D<T, default_value>::update_array_size(const Vector3 &size, bool reset) {
+        //Save old values
+        int old_size{internal_array.size()};
+        Vector<T> old_array{internal_array};
+        int old_width{width};
+        int old_depth{depth};
+
+        //Set new values
+        width = static_cast<int>(size.x);
+        depth = static_cast<int>(size.y);
+        height = static_cast<int>(size.z);
+        internal_array = Vector<T>();
+
+        for (int i = 0; i < compute_array_size(); i++) {
+            internal_array.push_back(default_value);
+        }
+
+        if (reset) {
+            return;
+        }
+
+        for (int i = 0; i < old_size; i++) {
+            int x = i % old_width;
+            int y = (i / old_width) % old_depth;
+            int z = i / (old_width * old_depth);
+
+            if (!is_in_range(x, y, z)) {
+                continue;
+            }
+
+            T value{old_array[i]};
+            if (value) {
+                set_data({static_cast<real_t>(x), static_cast<real_t>(y), static_cast<real_t>(z)}, value);
+            }
+        }
+    }
+
+    template<class T, T default_value>
+    void Grid3D<T, default_value>::reset() {
+        update_array_size({static_cast<real_t>(width), static_cast<real_t>(depth), static_cast<real_t>(height)}, true);
+    }
+
+    template<class T, T default_value>
+    T Grid3D<T, default_value>::get_data(const Vector3 &position) {
+        int pos_x = static_cast<int>(position.x);
+        int pos_y = static_cast<int>(position.y);
+        int pos_z = static_cast<int>(position.z);
+        return pos_x >= 0 && pos_x < width && pos_y >= 0 && pos_y < depth && pos_z >= 0 && pos_z < height
+               ? internal_array[pos_x + width * pos_y + width * depth * pos_z] : default_value;
+    }
+
+    template<class T, T default_value>
+    void Grid3D<T, default_value>::set_data(const Vector3 &position, T data) {
+        internal_array.set(
+                static_cast<int>(position.x) + width * static_cast<int>(position.y) +
+                width * depth * static_cast<int>(position.z),
+                data);
+    }
+
+    template<class T, T default_value>
+    bool Grid3D<T, default_value>::insert_box(const AABB &aabb, T data, bool remove) {
+        const Vector3 &position{aabb.position};
+        const Vector3 &size{aabb.size};
+        int index{get_index_from_position(position)};
+        int sizeX = static_cast<int>(size.x);
+        int sizeY = static_cast<int>(size.y);
+        int sizeZ = static_cast<int>(size.z);
+        int endX = static_cast<int>(position.x) + sizeX;
+        int endY = static_cast<int>(position.y) + sizeY;
+        int endZ = static_cast<int>(position.z) + sizeZ;
+
+        if (endX > width || endY > depth || endZ > height) {
+            return false;
+        }
+
+        internal_array.set(index, remove ? nullptr : data);
+        for (int i = 1; i < sizeX * sizeY * sizeZ; i++) {
+            index += Grid3D::index_increment_from(plane_square_and_jumps_from(size), size, i);
+            internal_array.set(index, remove ? nullptr : data);
+        }
+        return true;
+    }
+
+    template<class T, T default_value>
+    bool Grid3D<T, default_value>::is_overlapping(const AABB &aabb) const {
+        int index{get_index_from_position(aabb.position)};
+        const Vector3 &size{aabb.size};
+
+        if (index >= 0 && index < internal_array.size()) {
+            T element = internal_array[index];
+            if (element) {
+                return true;
+            }
+            for (int i = 1; i < static_cast<int>(size.x) * static_cast<int>(size.y) * static_cast<int>(size.z); i++) {
+                index += Grid3D::index_increment_from(plane_square_and_jumps_from(size), size, i);
+                if (index >= 0 && index < internal_array.size()) {
+                    element = internal_array[index];
+                    if (element) {
+                        return true;
+                    }
+                }
+            }
+        }
+        return false;
+    }
+
+    template<class T, T default_value>
+    bool Grid3D<T, default_value>::has(T object) const {
+        for (int i = 0; i < internal_array.size(); i++) {
+            if (internal_array[i] == object) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    template<class T, T default_value>
+    void Grid3D<T, default_value>::for_each(void (* fptr)(T)) {
+        for (int i = 0; i < internal_array.size(); i++) {
+            T element = internal_array[i];
+            if (element) {
+                fptr(element);
+            }
+        }
+    }
+
+    template<class T, T default_value>
+    int Grid3D<T, default_value>::get_width() const {
+        return width;
+    }
+
+    template<class T, T default_value>
+    void Grid3D<T, default_value>::set_width(int w) {
+        width = w;
+    }
+
+    template<class T, T default_value>
+    int Grid3D<T, default_value>::get_depth() const {
+        return depth;
+    }
+
+    template<class T, T default_value>
+    void Grid3D<T, default_value>::set_depth(int d) {
+        depth = d;
+    }
+
+    template<class T, T default_value>
+    int Grid3D<T, default_value>::get_height() const {
+        return height;
+    }
+
+    template<class T, T default_value>
+    void Grid3D<T, default_value>::set_height(int h) {
+        height = h;
+    }
+
+    template<class T, T default_value>
+    const Vector<T> &Grid3D<T, default_value>::get_internal_array() const {
+        return internal_array;
+    }
+
+    template<class T, T default_value>
+    void Grid3D<T, default_value>::set_internal_array(const Vector<T> &array) {
+        internal_array = array;
+    }
+
+    template<class T, T default_value>
+    bool Grid3D<T, default_value>::is_in_range(int x, int y, int z) const {
+        return x < width && y < depth && z < height;
+    }
+
+    template<class T, T default_value>
+    int Grid3D<T, default_value>::get_index_from_position(const Vector3 &position) const {
+        return static_cast<int>(position.x) + width * static_cast<int>(position.y) +
+               width * depth * static_cast<int>(position.z);
+    }
+
+    template<class T, T default_value>
+    Vector3 Grid3D<T, default_value>::get_position_3d(int id) const {
+        return {
+                static_cast<real_t>(id % width),
+                static_cast<real_t>((id / width) % depth),
+                static_cast<real_t>(id) / (width * depth)
+        };
+    }
+
+    template<class T, T default_value>
+    Vector3 Grid3D<T, default_value>::plane_square_and_jumps_from(const Vector3 &size) const {
+        return {
+                size.x * size.y,
+                static_cast<real_t>(width) - size.x,
+                width * (depth - size.y)
+        };
+    }
+
+    template<class T, T default_value>
+    int Grid3D<T, default_value>::index_increment_from(const Vector3 &planeSquareAndJumps, const Vector3 &size, int iteration) {
+        int increment{0};
+        increment += (iteration % static_cast<int>(size.x)) == 0 ? static_cast<int>(planeSquareAndJumps.y) : 0;
+        increment +=
+                (iteration % static_cast<int>(planeSquareAndJumps.x)) == 0 ? static_cast<int>(planeSquareAndJumps.z) : 0;
+        increment += 1;
+        return increment;
+    }
+
 } // namespace containers
 
 #endif //ISOMETRIC_MAPS_GRID_3D_H
